Tools/requirements.txt
google-cloud-storage
tqdm
Pillow